requirements.txt
streamlit
Pillow
streamlit-lottie
tensorflow
numpy
tensorflow.keras.models